Removed LocalService class and file. Added new common/api to handle execution of actions, reports, creating windows, etc. Removed spool and other Service related stuff. Documentation still to be written though.
Solved a bug in the way:
- Opening my requests filled the model group with double models. This was due to
the fact that we not only created a window with a given domain but we also told
the screen which ids had to go there. By passing False instead of the list of ids
we solved the problem and we reduce the number of RPC requests and thus improved
performance opening these tabs.