Removed a 'recordChanged()' signal:
It caused a strange crash on the following situation detected with spanish localization
modules that used an on_change trigger when the bank account of a partner is modified. When
the user modified that field and clicked save without the widget loosing focus first it'd 
crash with the following backtrace and message:

Traceback (most recent call last):
  File "/home/albert/d/koo/Koo/Dialogs/KooMainWindow.py", line 577, in callChildView
    res = wid.handlers[action]()
  File "/home/albert/d/koo/Koo/Dialogs/FormWidget.py", line 248, in save
    id = self.screen.save_current()
  File "/home/albert/prog/empresa/tinyerp/bzr/openobject-client-kde/Koo/Screen/Screen.py", line 433, in save_current
    self.current_view.store()
  File "/home/albert/d/koo/Koo/View/Form/FormView.py", line 158, in store
    self.widgets[name].store()
  File "/home/albert/d/koo/Koo/View/Form/one2many.py", line 167, in store
    self.screen.current_view.store()
  File "/home/albert/d/koo/Koo/View/Form/FormView.py", line 154, in store
    for name in self.widgets:
TypeError: Signal has 0 arguments, but 1 given

That 'Signal' refers to the 'emit' removed in this patch. It seems save to remove it as it is emited
anyway by the set() function that is called in the same function. That one doesn't throw the exception
which makes me think the problem might be in it being emited twice, though I'm not sure at all.