Use loadUiType() instead of loadUi for performance reasons. Loading complex forms can gain some hundreds of milliseconds (100-300). This is most important for form widgets but changed all UI's in the process.
Remove SupportDialog and action.

Changed type()==type() comparison for isinstance() it's not a major performance
gain at all but it's simply better.

Added Pyro protocol in Rpc, but no UI changes to choose it.

Use QImage to scale the image in an ImageFormWidget because performance is much
better.

Improved memory usage in binary and image widgets by not storing a copy of the 
field's data in memory. Store binaries in binary form (instead of base64) for
better memory usage too.