Revert revision 499. Refreshing in ViewForm is necessary due to onchange triggers. Solve the problem of fields that don't get stored in the model when buttons are pressed has to be solved in each button by now. So I've added a call to store the model in each button that can modify the model in ManyToOne, OneToMany and Reference widget. ManyToMany will follow in next commit.
Also fixed refreshing of calendar date as setting '1234' into a time widget it was converted
into '12:34:00' only the first time it was set (because the value had been modified). If you
wrote the text ('1234') again in the field and trigger lost focus, text was not converted into 
'12:34:00' because value had not changed.