Screen/Action: do NOT enable the arbitrary shortcuts by default
The idea of randomly assigning Ctrl+<0-9> etc. shortcuts to the actions,
which come at an arbitrary order from the server is fundamentally
flawed. In one recent incident, a stuck "Ctrl+" keypress from a barcode
scanner triggered unexpected actions!

Only if the user is really sure about that, he could enable that feature.