screen: don't put the active record's fields in context
the context of an action was filled with all data fields of the
current record. Apart from bloating the context, this introduced a
side-effect, when the 'journal_id' of account.analytic.line was
confused for an explicit 'account.journal.id'..

Shouldn't we pass these data, if ever needed, in a separate
"local context"?