Subscriber: do not attempt GUI at non-main thread
using notifications in the subscription sub-thread meant that a GUI
message would try to appear. In Qt, this would be a disaster.

We could have setup a signal/slot mechanism to emit those errors to the
main thread and alert the user. Who cares? Let them silently retry
to continue.