Rpc: provide `time` to safe_eval(), don't pollute `context`
Several context expressions, supplied by the server, contain refs
to time(), which is supposed to be in the client's evaluation globals.

However, we cannot put time() and `uid` inside the context, because
that is persistent and may be passed through RPC, too. So, put them
in the 'globals' dictionary, while the rest are in 'locals'.